move NGramKeyer test from KeyerTests to NGramFingerprintKeyerTest
