+++ main/tests/server/src/com/google/refine/clustering/binning/KeyerTests.java
@@ -122,19 +122,6 @@ public class KeyerTests extends RefineTest {
             { "", "" },
     };
 
-    private static final String[][] testNGramStrings = {
-            { "abcdefg", "abbccddeeffg" },
-            { "gfedcba", "bacbdcedfegf" },
-            { "a b c d e f g", "abbccddeeffg" },
-            { " a,b.c d\te!f?g ", "abbccddeeffg" },
-            { "écÉCec", "ceec" },
-            // All the whitespace characters below should be skipped
-            { "a\u0009\nb\u000Bc\u000Cd\re\u0085f\u00A0g\u1680h\u2000i\u2001j\u2002k\u2003l\u2004m\u2005n\u2006o\u2007p\u2008q\u2009r\u200As\u2028t\u2029u\u202Fv\u205Fw\u3000z",
-                    "abbccddeeffgghhiijjkkllmmnnooppqqrrssttuuvvwwz" },
-            { "", "" }, // TODO: add more test cases
-            { "", "" },
-    };
-
     @Override
     @BeforeTest
     public void init() {
@@ -159,16 +146,6 @@ public class KeyerTests extends RefineTest {
     @Test
     public void testFingerprintKeyer() {
         for (String[] ss : testStrings) {
-            Assert.assertEquals(ss.length, 2, "Invalid test"); // Not a valid test
-            Assert.assertEquals(keyer.key(ss[0]), ss[1],
-                    "Fingerprint for string: " + ss[0] + " failed");
-        }
-    }
-
-    @Test
-    public void testNGramKeyer() {
-        keyer = new NGramFingerprintKeyer();
-        for (String[] ss : testNGramStrings) {
             Assert.assertEquals(ss.length, 2, "Invalid test"); // Not a valid test
             Assert.assertEquals(keyer.key(ss[0]), ss[1],
                     "Fingerprint for string: " + ss[0] + " failed");
